🚚 Move common NFT code into mixins
